% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{compute.power}
\alias{compute.power}
\title{Compute power for single sample size replications.}
\usage{
compute.power(replications, target = 0.8)
}
\description{
Compute power for single sample size replications.
}
