% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exception.R
\name{Exception}
\alias{Exception}
\title{Package Exceptions}
\format{
\describe{
\item{\code{Exception$abstract_class_not_instantiable(object)}}{Exception for instantiating abstract classes or interfaces.}
\item{\code{Exception$method_not_implemented()}}{Exception for calling methods without an implementation.}
\item{\code{Exception$feature_not_developed()}}{Exception for running into things not yet developed.}
\item{\code{Exception$type_not_assignable(actual, expected)}}{Exception for when providing incorrect object types.}
\item{\code{Exception$unknown_package_option(option)}}{Exception for when requesting unknown package options.}
\item{\code{Exception$unnamed_argument_not_allowed()}}{Exception for providing unnamed arguments in the \code{...} construct.}
\item{\code{Exception$unknown_model_operation(operation)}}{Exception for calling an unknown model operation.}
}
}
\description{
This class contains static methods for throwing exceptions with informative
messages.
}

