% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionModelCreator.R
\name{FunctionModelCreator}
\alias{FunctionModelCreator}
\title{FunctionModelCreator}
\description{
This class is a concrete implementation of the \code{\link{ModelCreator}}
abstract class. This class is used internally by the framework to dynamically
create model implementations at runtime from provided functions for
specifying true model parameters, generating data, estimating model
parameters, and evaluating the estimated model parameters.
}
\seealso{
\code{\link{ModelCreator}}, \code{\link{ModelImplementation}}, and
\code{\link{RuntimeModel}}, and \code{\link{ModelService}}.
}
\section{Super class}{
\code{\link[powerly:ModelCreator]{powerly::ModelCreator}} -> \code{FunctionModelCreator}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FunctionModelCreator-new}{\code{FunctionModelCreator$new()}}
\item \href{#method-FunctionModelCreator-create_implementation}{\code{FunctionModelCreator$create_implementation()}}
\item \href{#method-FunctionModelCreator-clone}{\code{FunctionModelCreator$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="powerly" data-topic="ModelCreator" data-id="create"><a href='../../powerly/html/ModelCreator.html#method-ModelCreator-create'><code>powerly::ModelCreator$create()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionModelCreator-new"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionModelCreator-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{\link{FunctionModelCreator}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionModelCreator$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An object of class \code{\link{FunctionModelCreator}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionModelCreator-create_implementation"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionModelCreator-create_implementation}{}}}
\subsection{Method \code{create_implementation()}}{
Create a \code{\link{ModelImplementation}} instance from provided
functions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionModelCreator$create_implementation(
  specify,
  generate,
  estimate,
  evaluate
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{specify}}{A function that specifies the true model parameters.
The function must match the signature and return type of the
\code{specify} method in the \code{\link{ModelService}} class.}

\item{\code{generate}}{A function that generates a sample from the true
model. The function must match the signature and return type of the
\code{generate} method in the \code{\link{ModelService}} class.}

\item{\code{estimate}}{A function that estimates the model parameters. The
function must match the signature and return type of the \code{estimate}
method in the \code{\link{ModelService}} class.}

\item{\code{evaluate}}{A function that evaluates the estimation performance.
The function must match the signature and return type of the
\code{evaluate} method in the \code{\link{ModelService}} class.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{\link{ModelImplementation}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionModelCreator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionModelCreator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionModelCreator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
