% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{generate_model}
\alias{generate_model}
\title{Generate true model parameters}
\usage{
generate_model(type, ...)
}
\arguments{
\item{type}{Character string representing the type of true model. Possible
values are \code{"ggm"} (the default).}

\item{...}{Required arguments used for the generation of the true model. See
the \strong{True Models} section of \code{\link[=powerly]{powerly()}} for the arguments
required for each true model.}
}
\value{
A matrix containing the model parameters.
}
\description{
Generate matrices of true model parameters for the supported true models.
These matrices are intended to passed to the \code{model_matrix} argument of
\code{\link[=powerly]{powerly()}}.
}
\seealso{
\code{\link[=powerly]{powerly()}}, \code{\link[=validate]{validate()}}
}
