% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\docType{data}
\name{LOGO}
\alias{LOGO}
\title{The \code{powerly} package logo}
\format{
An object of class \code{character} of length 1.
}
\usage{
LOGO
}
\description{
The logo is generated by \code{\link[=make_logo]{make_logo()}} and displayed at package load time.
}
\keyword{internal}
