% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{make_logo}
\alias{make_logo}
\title{Update package logo displayed at load time.}
\usage{
make_logo(
  ascii.logo.path = "./inst/assets/logo/logo.txt",
  version = c(1, 0, 0)
)
}
\description{
This function is meant for generating or updating the logo.
After running this procedure we end up with what is stored in the \code{LOGO} constant.
}
