% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_1.R
\name{mc.run}
\alias{mc.run}
\title{Monte Carlo simulation.}
\usage{
mc.run(
  model,
  n,
  performance.measure,
  generate = ggm$generate,
  estimate = ggm$estimate,
  evaluate = ggm$evaluate,
  ...
)
}
\description{
Monte Carlo simulation.
}
