% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelConfiguration.R
\name{ModelConfiguration}
\alias{ModelConfiguration}
\title{ModelConfiguration}
\description{
This class is an implementation of the \code{\link{ModelConfigurationService}}
interface that provides a way to configure the arguments for the \code{specify},
\code{generate}, \code{estimate}, and \code{evaluate} operations defined by the
\code{\link{ModelService}} interface.
}
\seealso{
\code{\link{ModelConfigurationService}} and \code{\link{ModelService}}.
}
\section{Super class}{
\code{\link[powerly:ModelConfigurationService]{powerly::ModelConfigurationService}} -> \code{ModelConfiguration}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{specify}}{A list of arguments for the \code{specify} operation.}

\item{\code{generate}}{A list of arguments for the \code{generate} operation.}

\item{\code{estimate}}{A list of arguments for the \code{estimate} operation.}

\item{\code{evaluate}}{A list of arguments for the \code{evaluate} operation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelConfiguration-new}{\code{ModelConfiguration$new()}}
\item \href{#method-ModelConfiguration-configure_specify}{\code{ModelConfiguration$configure_specify()}}
\item \href{#method-ModelConfiguration-configure_generate}{\code{ModelConfiguration$configure_generate()}}
\item \href{#method-ModelConfiguration-configure_estimate}{\code{ModelConfiguration$configure_estimate()}}
\item \href{#method-ModelConfiguration-configure_evaluate}{\code{ModelConfiguration$configure_evaluate()}}
\item \href{#method-ModelConfiguration-clone}{\code{ModelConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{\link{ModelConfiguration}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfiguration$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An object of class \code{\link{ModelConfiguration}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfiguration-configure_specify"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfiguration-configure_specify}{}}}
\subsection{Method \code{configure_specify()}}{
Configure the arguments for the \code{specify} operation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfiguration$configure_specify(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments for the \code{specify} operation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This method returns void. The updated list of arguments can be
accessed via the \code{specify} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfiguration-configure_generate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfiguration-configure_generate}{}}}
\subsection{Method \code{configure_generate()}}{
Configure the arguments for the \code{generate} operation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfiguration$configure_generate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments for the \code{generate} operation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This method returns void. The updated list of arguments can be
accessed via the \code{generate} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfiguration-configure_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfiguration-configure_estimate}{}}}
\subsection{Method \code{configure_estimate()}}{
Configure the arguments for the \code{estimate} operation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfiguration$configure_estimate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments for the \code{estimate} operation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This method returns void. The updated list of arguments can be
accessed via the \code{estimate} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfiguration-configure_evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfiguration-configure_evaluate}{}}}
\subsection{Method \code{configure_evaluate()}}{
Configure the arguments for the \code{evaluate} operation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfiguration$configure_evaluate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments for the \code{evaluate} operation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This method returns void. The updated list of arguments can be
accessed via the \code{evaluate} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
