% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelConfigurationService.R
\name{ModelConfigurationService}
\alias{ModelConfigurationService}
\title{ModelConfigurationService}
\description{
This is an interface that defines the methods required for configuring
arguments passed to concrete model implementations of the
\code{\link{ModelService}} interface.
}
\seealso{
\code{\link{ModelConfiguration}} and \code{\link{ModelService}}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelConfigurationService-new}{\code{ModelConfigurationService$new()}}
\item \href{#method-ModelConfigurationService-configure_specify}{\code{ModelConfigurationService$configure_specify()}}
\item \href{#method-ModelConfigurationService-configure_generate}{\code{ModelConfigurationService$configure_generate()}}
\item \href{#method-ModelConfigurationService-configure_estimate}{\code{ModelConfigurationService$configure_estimate()}}
\item \href{#method-ModelConfigurationService-configure_evaluate}{\code{ModelConfigurationService$configure_evaluate()}}
\item \href{#method-ModelConfigurationService-clone}{\code{ModelConfigurationService$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfigurationService-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfigurationService-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{\link{ModelConfigurationService}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfigurationService$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Instantiating this class will throw an error.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfigurationService-configure_specify"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfigurationService-configure_specify}{}}}
\subsection{Method \code{configure_specify()}}{
Configure the arguments for the \code{specify} operation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfigurationService$configure_specify(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments for the \code{specify} operation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This method returns void.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfigurationService-configure_generate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfigurationService-configure_generate}{}}}
\subsection{Method \code{configure_generate()}}{
Configure the arguments for the \code{generate} operation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfigurationService$configure_generate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments for the \code{generate} operation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This method returns void.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfigurationService-configure_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfigurationService-configure_estimate}{}}}
\subsection{Method \code{configure_estimate()}}{
Configure the arguments for the \code{estimate} operation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfigurationService$configure_estimate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments for the \code{estimate} operation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This method returns void.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfigurationService-configure_evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfigurationService-configure_evaluate}{}}}
\subsection{Method \code{configure_evaluate()}}{
Configure the arguments for the \code{evaluate} operation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfigurationService$configure_evaluate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments for the \code{evaluate} operation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This method returns void.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfigurationService-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfigurationService-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfigurationService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
