% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCreator.R
\name{ModelCreator}
\alias{ModelCreator}
\title{ModelCreator}
\description{
This is an abstract class that defines the algorithm for creating a
\code{\link{RuntimeModel}} instance from a \code{\link{ModelImplementation}}
instance. Concrete implementations of this class must handle the creation of
\code{\link{ModelImplementation}} instances (e.g., from a set of functions
provided in the global environment). For registering a new model with the
framework, it is recommended to use use the public \code{API} facades
\code{\link{register_model}} and \code{\link{validate_model}}.
}
\seealso{
\code{\link{ModelService}}, \code{\link{ModelImplementation}},
\code{\link{RuntimeModel}}, and \code{\link{FunctionModelCreator}}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelCreator-new}{\code{ModelCreator$new()}}
\item \href{#method-ModelCreator-create_implementation}{\code{ModelCreator$create_implementation()}}
\item \href{#method-ModelCreator-create}{\code{ModelCreator$create()}}
\item \href{#method-ModelCreator-clone}{\code{ModelCreator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelCreator-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelCreator-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{\link{ModelCreator}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCreator$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Instantiating this class will throw an error.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelCreator-create_implementation"></a>}}
\if{latex}{\out{\hypertarget{method-ModelCreator-create_implementation}{}}}
\subsection{Method \code{create_implementation()}}{
An abstract method for creating a \code{\link{ModelImplementation}}
instance from a set of arguments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCreator$create_implementation(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed from the global environment for creating
a \code{\link{ModelImplementation}} instance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{\link{ModelImplementation}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelCreator-create"></a>}}
\if{latex}{\out{\hypertarget{method-ModelCreator-create}{}}}
\subsection{Method \code{create()}}{
Create a \code{\link{RuntimeModel}} instance from a set of arguments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCreator$create(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed from the global environment that are
provided to the \code{create_implementation} method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{\link{RuntimeModel}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelCreator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelCreator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCreator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
