% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelImplementation.R
\name{ModelImplementation}
\alias{ModelImplementation}
\title{ModelImplementation}
\description{
An intermediary class that holds the implementation details for a statistical
model. This class is used internally by the framework to dynamically create
model implementations at runtime.
}
\seealso{
\code{\link{ModelService}}, \code{\link{RuntimeModel}},
\code{\link{ModelCreator}}, and \code{\link{FunctionModelCreator}}.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{specify}}{The provided function for specifying the true model
parameters.}

\item{\code{generate}}{The provided function for generating a sample from
the true model parameters.}

\item{\code{estimate}}{The provided function for estimating the model
parameters.}

\item{\code{evaluate}}{The provided function for evaluating the estimation
performance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelImplementation-set_specify}{\code{ModelImplementation$set_specify()}}
\item \href{#method-ModelImplementation-set_generate}{\code{ModelImplementation$set_generate()}}
\item \href{#method-ModelImplementation-set_estimate}{\code{ModelImplementation$set_estimate()}}
\item \href{#method-ModelImplementation-set_evaluate}{\code{ModelImplementation$set_evaluate()}}
\item \href{#method-ModelImplementation-clone}{\code{ModelImplementation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelImplementation-set_specify"></a>}}
\if{latex}{\out{\hypertarget{method-ModelImplementation-set_specify}{}}}
\subsection{Method \code{set_specify()}}{
Set the function for specifying the true model parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelImplementation$set_specify(specify)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{specify}}{A function that specifies the true model parameters.
The function must match the signature and return type of the
\code{specify} method in the \verb{[powerly::ModelService]} class.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This method returns void.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelImplementation-set_generate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelImplementation-set_generate}{}}}
\subsection{Method \code{set_generate()}}{
Set the function for generating a sample from the true model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelImplementation$set_generate(generate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{generate}}{A function that generates a sample from the true
model. The function must match the signature and return type of the
\code{generate} method in the \verb{[powerly::ModelService]} class.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This method returns void.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelImplementation-set_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelImplementation-set_estimate}{}}}
\subsection{Method \code{set_estimate()}}{
Set the function for estimating the model parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelImplementation$set_estimate(estimate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{estimate}}{A function that estimates the model parameters. The
function must match the signature and return type of the \code{estimate}
method in the \verb{[powerly::ModelService]} class.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This method returns void.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelImplementation-set_evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelImplementation-set_evaluate}{}}}
\subsection{Method \code{set_evaluate()}}{
Set the function for evaluating the estimation performance by a
performance measure of interest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelImplementation$set_evaluate(evaluate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{evaluate}}{A function that evaluates the estimation performance.
The function must match the signature and return type of the
\code{evaluate} method in the \verb{[powerly::ModelService]} class.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This method returns void.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelImplementation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelImplementation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelImplementation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
