% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelService.R
\name{ModelService}
\alias{ModelService}
\title{ModelService}
\description{
This is an interface that defines the operations a statistical model must
implement in order to be receive sample size analysis support. Models are
created and handled internally by the framework. For registering a new model
with the framework, please use the public \code{API} \code{\link{register_model}}
and \code{\link{validate_model}}.
}
\seealso{
\code{\link{RuntimeModel}}, \code{\link{ModelImplementation}},
\code{\link{ModelCreator}}, \code{\link{FunctionModelCreator}},
\code{\link{register_model}}, and \code{\link{validate_model}}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelService-new}{\code{ModelService$new()}}
\item \href{#method-ModelService-specify}{\code{ModelService$specify()}}
\item \href{#method-ModelService-generate}{\code{ModelService$generate()}}
\item \href{#method-ModelService-estimate}{\code{ModelService$estimate()}}
\item \href{#method-ModelService-evaluate}{\code{ModelService$evaluate()}}
\item \href{#method-ModelService-clone}{\code{ModelService$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelService-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelService-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{\link{ModelService}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelService$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Instantiating this class will throw an error.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelService-specify"></a>}}
\if{latex}{\out{\hypertarget{method-ModelService-specify}{}}}
\subsection{Method \code{specify()}}{
Specify the true (i.e., population) model parameters. The output of
this method is used as input for the \code{generate} and \code{evaluate}
methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelService$specify(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments passed from the global environment for
creating a true model specification (i.e., the true model
parameters).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of variable length, containing the results of the model
specification logic.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelService-generate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelService-generate}{}}}
\subsection{Method \code{generate()}}{
Generate a sample (i.e., data) from the true model specification.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelService$generate(specification, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{specification}}{A list of variable length, representing the
output of the \code{specify} method.}

\item{\code{...}}{Optional arguments passed from the global environment for
generating a sample based on the true model specification.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of variable length, containing the generated sample.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelService-estimate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelService-estimate}{}}}
\subsection{Method \code{estimate()}}{
Estimate the model parameters from the generated sample. The output
of this method serves as part of the input for the \code{evaluate} method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelService$estimate(generation, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{generation}}{A list of variable length, representing the output
of the \code{generate} method.}

\item{\code{...}}{Optional arguments passed from the global environment
required for estimating the model parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of variable length, containing the estimated model parameters.
The output of this method serves as part of the input for the
\code{evaluate} method.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelService-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelService-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
Compute a performance measure of interest by comparing the true model
parameters with the estimated model parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelService$evaluate(specification, estimation, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{specification}}{A list of variable length, representing the
output of the \code{specify} method.}

\item{\code{estimation}}{A list of variable length, representing the output
of the \code{estimate} method.}

\item{\code{...}}{Optional arguments passed from the global environment
required for computing the performance measure.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of variable length, containing the performance measure of
interest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelService-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelService-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
