% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSimulation.R
\name{ModelSimulation}
\alias{ModelSimulation}
\title{ModelSimulation}
\description{
This class implements the \code{\link{SimulationService}} interface for
running a simulation using a \code{\link{ModelService}} implementation
as a backend.
}
\seealso{
\code{\link{SimulationService}}, \code{\link{ModelService}}, and
\code{\link{ModelConfiguration}}.
}
\section{Super class}{
\code{\link[powerly:SimulationService]{powerly::SimulationService}} -> \code{ModelSimulation}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{An object of a class implementing the
\code{\link{ModelService}} interface.}

\item{\code{configuration}}{An object of a class implementing the
\code{\link{ModelConfigurationService}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelSimulation-new}{\code{ModelSimulation$new()}}
\item \href{#method-ModelSimulation-configure}{\code{ModelSimulation$configure()}}
\item \href{#method-ModelSimulation-run}{\code{ModelSimulation$run()}}
\item \href{#method-ModelSimulation-clone}{\code{ModelSimulation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelSimulation-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelSimulation-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{\link{ModelSimulation}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelSimulation$new(model, configuration)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{An object of a class implementing the
\code{\link{ModelService}} interface.}

\item{\code{configuration}}{An object of a class implementing the
\code{\link{ModelConfigurationService}} interface.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{\link{ModelSimulation}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelSimulation-configure"></a>}}
\if{latex}{\out{\hypertarget{method-ModelSimulation-configure}{}}}
\subsection{Method \code{configure()}}{
Configure the arguments for a specific model operation defined by the
\code{\link{ModelService}} interface.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelSimulation$configure(operation, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{operation}}{A character string specifying the model operation to
configure.}

\item{\code{...}}{The arguments that will be passed to the model operation
when called during the simulation. The arguments are specific to the
model operation being configured and to the specific model
implementation. Registering a new model with the framework can be
done through the public \code{API} \code{\link{register_model}} and
\code{\link{validate_model}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This method returns void. The updated list of arguments can be
accessed via the appropriate field in the \code{configuration} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelSimulation-run"></a>}}
\if{latex}{\out{\hypertarget{method-ModelSimulation-run}{}}}
\subsection{Method \code{run()}}{
Run the simulation using the specific model and configuration
implementations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelSimulation$run(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments controlling the simulation flow. Note
that these are not passed to the model operations (i.e., see the
\code{configure} method for that). These arguments are currently not in
use.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This method returns a list of variable length, containing the results
of a single simulation run.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelSimulation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelSimulation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelSimulation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
