% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Options.R
\name{Options}
\alias{Options}
\title{Class for Package Options}
\description{
This class holds public fields that represent the package
\code{\link[base:options]{options}} used to configure the default behavior of the
functionality \code{\link{powerly}} provides.
}
\details{
An instance of this class is automatically created and stored in the session
\code{\link[base:options]{base::.Options}} at load time. This instance can be accessed and changed
via \code{\link[base:options]{getOption("powerly")}}. Specific package
\code{\link[base:options]{options}} can be retrieved using the helper function
\code{\link[=get_option]{get_option()}}.
}
\examples{
# Set the default package options (i.e., automatically set at load time).
set_default_options()

# ...

# Restore the defaults.
set_default_options()

}
\seealso{
\code{\link[=get_option]{get_option()}}, \code{\link[=set_option]{set_option()}}, and
\code{\link[=set_default_options]{set_default_options()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{verbose}}{A logical value indicating whether progress messages}
}
\if{html}{\out{</div>}}
}
