% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_settings}
\alias{plot_settings}
\title{Fetch \code{ggplot2} theme settings.}
\usage{
plot_settings()
}
\value{
A list of \code{ggplot2} layers.
}
\description{
This function is used to create reusable \code{ggplot2} theme settings.
}
\keyword{internal}
