% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerly-package.R
\docType{package}
\name{powerly-package}
\alias{powerly-package}
\alias{_PACKAGE}
\title{Sample Size Analysis Framework}
\description{
...
}
\details{
...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://powerly.dev}
  \item Report bugs at \url{https://github.com/mihaiconstantin/powerly/issues}
}

}
\author{
\strong{Maintainer}: Mihai Constantin \email{mihai@mihaiconstantin.com} (\href{https://orcid.org/0000-0002-6460-0107}{ORCID})

}
\keyword{internal}
