% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{run.method}
\alias{run.method}
\title{Run method.}
\usage{
run.method(
  model,
  range,
  replications,
  measure = "sen",
  target = 0.8,
  statistic = "power",
  criterion = 0.8,
  n.samples = 10,
  tolerance = 50,
  boots = 10000,
  monotone = TRUE,
  non.increasing = FALSE,
  runs = 10,
  ...,
  verbose = TRUE
)
}
\description{
Run method.
}
