% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_1.R
\name{run.step.1}
\alias{run.step.1}
\title{Step 1.}
\usage{
run.step.1(
  selected.sample.sizes,
  replications,
  performance.measure = "sen",
  performance.measure.target = 0.8,
  statistic.definition = "power",
  statistic.criterion = 0.8,
  ...,
  verbose = TRUE
)
}
\description{
Step 1.
}
