% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuntimeModel.R
\name{RuntimeModel}
\alias{RuntimeModel}
\title{RuntimeModel}
\description{
This class is a concrete implementation of the \code{\link{ModelService}}
interface. It is used to dynamically create model implementations at runtime
that can be used to specify true model parameters, generate data, estimate
model parameters, and evaluate the estimated model parameters (i.e, by
computing performance measures of interest). The runtime model is dispatching
method calls to a \code{\link{ModelImplementation}} instance.
}
\seealso{
\code{\link{ModelService}}, \code{\link{ModelImplementation}},
\code{\link{ModelCreator}}, and \code{\link{FunctionModelCreator}}.
}
\section{Super class}{
\code{\link[powerly:ModelService]{powerly::ModelService}} -> \code{RuntimeModel}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{implementation}}{The \code{\link{ModelImplementation}} instance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RuntimeModel-new}{\code{RuntimeModel$new()}}
\item \href{#method-RuntimeModel-set_implementation}{\code{RuntimeModel$set_implementation()}}
\item \href{#method-RuntimeModel-specify}{\code{RuntimeModel$specify()}}
\item \href{#method-RuntimeModel-generate}{\code{RuntimeModel$generate()}}
\item \href{#method-RuntimeModel-estimate}{\code{RuntimeModel$estimate()}}
\item \href{#method-RuntimeModel-evaluate}{\code{RuntimeModel$evaluate()}}
\item \href{#method-RuntimeModel-clone}{\code{RuntimeModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RuntimeModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-RuntimeModel-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{\link{RuntimeModel}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RuntimeModel$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An object of class \code{\link{RuntimeModel}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RuntimeModel-set_implementation"></a>}}
\if{latex}{\out{\hypertarget{method-RuntimeModel-set_implementation}{}}}
\subsection{Method \code{set_implementation()}}{
Set the \code{\link{ModelImplementation}} instance to be used by the
\code{\link{RuntimeModel}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RuntimeModel$set_implementation(implementation)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{implementation}}{An object of class
\code{\link{ModelImplementation}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This method returns void.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RuntimeModel-specify"></a>}}
\if{latex}{\out{\hypertarget{method-RuntimeModel-specify}{}}}
\subsection{Method \code{specify()}}{
Specify the true (i.e., population) model parameters. The output of
this method is used as input for the \code{generate} and \code{evaluate}
methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RuntimeModel$specify(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments passed from the global environment for
creating a true model specification (i.e., the true model
parameters).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of variable length, containing the results of the model
specification logic.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RuntimeModel-generate"></a>}}
\if{latex}{\out{\hypertarget{method-RuntimeModel-generate}{}}}
\subsection{Method \code{generate()}}{
Generate a sample (i.e., data) from the true model specification.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RuntimeModel$generate(specification, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{specification}}{A list of variable length, representing the
output of the \code{specify} method.}

\item{\code{...}}{Optional arguments passed from the global environment for
generating a sample based on the true model specification.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of variable length, containing the generated sample.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RuntimeModel-estimate"></a>}}
\if{latex}{\out{\hypertarget{method-RuntimeModel-estimate}{}}}
\subsection{Method \code{estimate()}}{
Estimate the model parameters from the generated sample. The output
of this method serves as part of the input for the \code{evaluate} method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RuntimeModel$estimate(generation, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{generation}}{A list of variable length, representing the output
of the \code{generate} method.}

\item{\code{...}}{Optional arguments passed from the global environment for
estimating the model parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of variable length, containing the estimated model parameters.
The output of this method serves as part of the input for the
\code{evaluate} method.
Compute a performance measure of interest by comparing the true model
parameters with the estimated model parameters.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RuntimeModel-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-RuntimeModel-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RuntimeModel$evaluate(specification, estimation, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{specification}}{A list of variable length, representing the
output of the \code{specify} method.}

\item{\code{estimation}}{A list of variable length, representing the output
of the \code{estimate} method.}

\item{\code{...}}{Optional arguments passed from the global environment for
computing the performance measure.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of variable length, containing the performance measure
of interest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RuntimeModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RuntimeModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RuntimeModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
