% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationService.R
\name{SimulationService}
\alias{SimulationService}
\title{SimulationService}
\description{
This is an interface that defines the methods a simulation class must
implement.
}
\seealso{
\code{\link{ModelSimulation}}, \code{\link{ModelService}}, and
\code{\link{ModelConfigurationService}}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SimulationService-new}{\code{SimulationService$new()}}
\item \href{#method-SimulationService-configure}{\code{SimulationService$configure()}}
\item \href{#method-SimulationService-run}{\code{SimulationService$run()}}
\item \href{#method-SimulationService-clone}{\code{SimulationService$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationService-new"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationService-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{\link{SimulationService}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationService$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Instantiating this class will throw an error.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationService-configure"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationService-configure}{}}}
\subsection{Method \code{configure()}}{
Configure the simulation arguments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationService$configure(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments for configuring the simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This method returns void.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationService-run"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationService-run}{}}}
\subsection{Method \code{run()}}{
Run the simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationService$run(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments for running the simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This method returns a list of variable length, containing the results
of a single simulation run.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationService-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationService-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
