% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_2.R
\name{spline.methdology}
\alias{spline.methdology}
\title{Apply spline methdology.}
\usage{
spline.methdology(x, y, inner.knots, monotone = TRUE, non.increasing = FALSE)
}
\description{
Apply spline methdology.
}
