% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splower-package.R
\docType{package}
\name{splower}
\alias{splower}
\alias{splower-package}
\title{\code{splower}: Sample Size Analysis for Psychological Network Models.}
\description{
The \code{splower} Run sample size analysis for cross-sectional network models.
}
\details{
The package is under active development.
First major release is expected on 01.31.2021.
}
\author{
\strong{Maintainer}: Mihai Constantin \email{constantin.mihai.alexandru@gmail.com} (\href{https://orcid.org/0000-0002-6460-0107}{ORCID})

}
