% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.Method}
\alias{summary.Method}
\title{Provide a summary of the sample size analysis results}
\usage{
\method{summary}{Method}(object, ...)
}
\arguments{
\item{object}{An object instance of class \code{Method}.}

\item{...}{Other optional arguments currently not in use.}
}
\description{
This function summarizes the objects of class \code{Method} and provides
information about the method run and the sample size recommendation.
}
\seealso{
\code{\link[=plot.Method]{plot.Method()}}, \code{\link[=powerly]{powerly()}}
}
