% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate.recommendation}
\alias{validate.recommendation}
\title{Validate recommendation.}
\usage{
validate.recommendation(
  model,
  recommendation,
  replications = 100,
  performance.measure = "sen",
  performance.measure.target = 0.8,
  statistic.definition = "power",
  ...,
  verbose = TRUE
)
}
\description{
Validate recommendation.
}
