% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Warning.R
\name{Warning}
\alias{Warning}
\title{Package Warnings}
\format{
\describe{
\item{\code{Warning$duplicate_arguments(names)}}{Warning for providing duplicate arguments in the \code{...} construct.}
}
}
\description{
This class contains static methods for throwing warnings with informative
messages.
}

