% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerly-package.R
\docType{package}
\name{powerly-package}
\alias{powerly-package}
\alias{_PACKAGE}
\title{Sample Size Analysis for Psychological Networks and More}
\description{
\code{powerly} is a package that implements the method by \href{https://arxiv.org}{Constantin et al. (2021)} for conducting sample size analysis for
network models.
}
\details{
The method implemented is implemented in the main function \code{\link[=powerly]{powerly()}}. The
implementation takes the form of a three-step recursive algorithm designed to
find an optimal sample size value given a model specification and an outcome
measure of interest. It starts with a Monte Carlo simulation step for
computing the outcome at various sample sizes. It continues with a monotone
curve-fitting step for interpolating the outcome. The final step employs
stratified bootstrapping to quantify the uncertainty around the fitted curve.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mihaiconstantin/powerly}
  \item Report bugs at \url{https://github.com/mihaiconstantin/powerly/issues}
}

}
\author{
\strong{Maintainer}: Mihai Constantin \email{mihai@mihaiconstantin.com} (\href{https://orcid.org/0000-0002-6460-0107}{ORCID})

}
\keyword{internal}
