% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method.R
\name{summary.Method}
\alias{summary.Method}
\title{Provide a summary of the results}
\usage{
\method{summary}{Method}(object, ...)
}
\arguments{
\item{object}{An object instance of class \code{Method}.}
}
\description{
This function summarizes the objects of class \code{Method} and provides
information about the method run and the sample size recommendation.
}
\keyword{internal}
