% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation.R
\name{summary.Validation}
\alias{summary.Validation}
\title{Provide a summary of the validation results}
\usage{
\method{summary}{Validation}(object, ...)
}
\arguments{
\item{object}{An object instance of class \code{Validation}.}
}
\description{
This function summarizes the objects of class \code{Validation} and provides
information.
}
\keyword{internal}
